refactor: add dashboard wrapper for my-app
