+++ packages/db-dashboard/src/pages/MyApp.jsx
@@ -1,25 +1,8 @@
 'use strict'
-import logo from '../../../my-app/src/logo.svg';
-import '../../../my-app/src/App.css';
+import MyApp from '../../../my-app/src/App';
 
 export default function App(props) {
-  const { swaggerDocUrl } = props
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
+    <MyApp/>
+  )
 }
