+++ packages/my-app/src/App.jsx
